requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.25
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
python-dotenv>=1.0.0
pymysql>=1.1.0

